version 0.3 rc2
git-svn-id: https://bitcoin.svn.sourceforge.net/svnroot/bitcoin/trunk@95 1a98c847-1fd6-4fd8-948a-caf3550aa51b

+++ main.cpp
@@ -2707,7 +2707,7 @@ void BitcoinMiner()
             }
 
             // Update nTime every few seconds
-            const uint64 nMask = 0xffff;
+            const unsigned int nMask = 0xffff;
             if ((++tmp.block.nNonce & nMask) == 0)
             {
                 // Meter hashes/sec
